+++ src/Responses/Traits/Interpreter.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Coreproc\PaynamicsSdk\Responses\Traits;
+
+trait Interpreter
+{
+    public function interpret()
+    {
+        //
+    }
+}
