+++ src/Services/QueryService.php
@@ -3,7 +3,7 @@
 namespace Coreproc\PaynamicsSdk\Services;
 
 use Coreproc\PaynamicsSdk\PaynamicsClient;
-use Coreproc\PaynamicsSdk\Request\QueryRequest;
+use Coreproc\PaynamicsSdk\Requests\QueryRequest;
 use Coreproc\PaynamicsSdk\Services\Interfaces\RequestInterface;
 use Coreproc\PaynamicsSdk\Services\Traits\GenerateForm;
 use SimpleXMLElement;
